moving selection helpers to a module
